[FIX]:Video id type 변경(Integer->Long)

+++ src/main/java/com/semtleWebGroup/youtubeclone/domain/video/repository/VideoRepository.java
@@ -3,7 +3,7 @@ import com.semtleWebGroup.youtubeclone.domain.video.domain.Video;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface VideoRepository extends JpaRepository<Video, Integer> {
+public interface VideoRepository extends JpaRepository<Video, Long> {
     List<Video> findByTitleContaining(String key);
 
 }
